clean up dev artifacts, move helper functions to utils file
